backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
web3>=6.0.0
python-dotenv>=0.19.0
beanie>=1.11.0
motor>=3.0.0
pymongo>=4.0.0
pydantic>=1.9.0
httpx>=0.23.0
requests>=2.27.0
python-multipart>=0.0.5
aiofiles>=0.8.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pytest>=7.0.0
pytest-asyncio>=0.18.0
black>=22.1.0
isort>=5.10.0
pylint>=2.12.0
Jinja2>=3.0.0
pyyaml>=6.0